WIP: start porting this to hugo
Hugo is a simple static site generator. These pages are mainly about
documentation and thus, are rather static in nature. Using a static
site generator instead of a single page application, allows for better
SEO, editing experience and results in a plethora of features, like
RSS feeds out of the box.